backend/requirements.txt
# API Framework
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2

# AI and LLM
langchain>=0.0.335
langchain-groq>=0.0.1
langchain-core>=0.1.0

# Web Search
tavily-python>=0.2.8

# Data Processing
yfinance>=0.2.31
matplotlib>=3.8.0
numpy>=1.26.0
pandas>=2.1.1

# Document Generation
reportlab>=4.0.4
python-docx>=1.0.0
markdown>=3.5.1
Pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
certifi>=2023.7.22